Keep refactoring `fractal_server/tasks`
